fix debian 13 boot start error: version OPENSSL_3.4.0 not found (#250)
When installing version 10.0.2 on Debian 13, the role throws the following error whenever enabling or disabling boot-start:
```
systemctl: /opt/splunk/lib/libcrypto.so.3: version `OPENSSL_3.4.0' not found (required by /usr/lib/x86_64-linux-gnu/systemd/libsystemd-shared-257.so)
systemctl: /opt/splunk/lib/libcrypto.so.3: version `OPENSSL_3.4.0' not found (required by /usr/lib/x86_64-linux-gnu/systemd/libsystemd-shared-257.so)
```
The actual exit status is 8 and the boot is actually configured correctly, but since it's not 0, the task fails.

The current solution is to limit that task failure when it's not 0 or 8.